requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
celery==5.4.0
redis==5.0.8
sqlmodel==0.0.22
psycopg2-binary==2.9.9
requests==2.32.3
pydantic==2.9.2
alembic==1.13.2
# Test dependencies
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
pytest-mock==3.12.0
requests-mock==1.11.0
python-dotenv==1.0.0
